xbman-Integrated/requirements.txt
Django==1.8.3
paramiko==2.0.2
python-jenkins==0.4.13
ansible==1.9.6
celery==3.1.17
django-celery==3.1.17
gunicorn